Implementation of WeaponCard + Report Sonar 04-21
Implementation of "WeaponCard" class. Minor changes on other classe. Bug fix.

For the implementation it use lambda function to associate the effect to the card and read JSON files to handle the weapon effect instead of using 21 class for each weapon (which have been deleted). In this way the code result more light and be scalable.

Vision of cards (the player you can actually shoot at) is demanded to "Game" class which read the enum associated to a weapon card and convert it to a proper function.

Delete method to handle the advanced effect because we have choosen to write the simple rules and upgrade them after the whole basic game works.

Add modelView packages with classes skeleton to handle the data. This classes are used for data flowing from model to view.

Fix a bug where a PowerUpEffect implement the wrong interface.