Implementation of Square and it's subclasses
Square class now use 4 boolean to check if there are walls at the cardinal position of a square and assign 4 Square variable, instead of the enum InfoCardinal, in order to know which room surronds it.
It now use an idRoom to identify the room it belongs to, because color was too generic.

Add:
- Value in AreaOfEffect to manage the distance
Removed:
- Enum InfoCardinal now deprecated