Implementation of AmmoCard, PowerUpCard with their effect, Card, Deck and DeckFactory classes.
Added DeckFactory instead of CardFactory to generate a deck of  card (list) and not a single card.

Removed method 'effect' in PowerUp class, because is redundant.

Added a new method "drawPowerUp" from Deck to draw a powerUp and add to player powerUps list.

Added a new package and class, "IdGenerator", to generate an incremental id to identify the cards.

Fix bug constructor WeaponCard

Report sonar 04-28