Added package Controller and classes.
Added LightModel and Session classes in the Model package.

Added TestDeck Test PowerUp classes implementation.

Fixed AmmoType classes with a int value associated to the enumType; easier to link the color with the Square.

Added a first implementation of SpawnController.

Added a new method to TagbackGrenadeEff to check if the powerUp can be used.