Add new tests and implement new methods for scoring
Add new tests for the player and working tests for turn manager and game table.

Almost completed implementation of all the methods concerning scoring and added all the methods needed (even if two are still not implemented).

Consider getNumberOfPlayers as deprecated. Instead of using GameTable.getNumberOfPlayers() we can use GameTable.getPlayers().length.
getNumberOfPLayers is still in the code to avoid problems with team mates. It will be removed later.