Updated network and fixed minor bug
Network handler and Virtual View are now reduced to one class each. In this way the View client side can communicate with server through them in a more clean way.

In order to avoid to check the instances of every message class, the messages are now incorporeted in an Event class which also have an attribute to differentiate the type of event. This would be useful for the events' observers because they could know if there is something of their competence to be handled after a notify.

Make order of packages and refactor the reference about them.

With this structure the MVC pattern is preserved and the network is totally hidden.