LOGIN MANAGEMENT - USER INPUT MANAGEMENT
Add class to handle user's login.
When a client connect, the server create a new session and keep track of the client's that has been connected.
This way, whenever an user make an action, the server can check if he is authenticated. Moreover, when user disconnect, he can reconnect if he use the same username and the same pc to come back.

Add class to handle input from user.
It get user input using a different thread. This way, whenever an error occur or a timer goes off, it can be interupted and stop the user from sending new data.

Bug fix:
- In IdConverter the card are now retrived correctly