DistanceDictionary revision + Bug fix
Distance dictionary now adopt a strategy pattern. In this way the code is more clear than having a lot of function in one class. Moreover, the method which return the aviable position in this class is now static in order to have access to it without having to pass the dictionary throughout the whole project when needed.

Some bug fixes:
- InputScanner don't reset the timer for action whenever an user do a wrong choice.