Add library for colors and update GeneralInfo + others
The library jansi has been added.

Now the ids of he different rooms (also used to determine the color) are in GeneralInfo.

Implemented part of the Reload State in the View.

moveTo in Player has been updated as requested by Luca Iovine, to move also the reference of the player from a square to another, instead of just changing the reference of the square on the player.

The problem in the test of hasEnoughAmmo in the TestPlayer has been fixed.