Implementation of Weapon Effect
The weapon effect are fully implemented and tested.
The access to the neccessary data is done through the WeaponCard class.
There are some specific class of the weapon effect which couldn't be more generic.

From the weapon effect class could be retrived the possible enemy hittable, where an enemy could be moved after the effect and the check of the choice made.

Update the Json file for the weapon effect configuration.

Change two type of AreaOfEffect because they were wrong.

Add new WeaponEffectTYpe