Update view + fix inconsistencies in gameTable model and tests
LoginState, SetupGameState and ReloadState have been partially implemented.

The interface IState has been deprecated and it has been substituted by the abstract class State, to make it able to have some methods in common with all the states without losing any of the previous advantages of the interface.

The overall of the view has been changed with the creation of observers that are not states and also an abstract class called ToolsView: both used by states for methods of common use.

The representation of the map in InfoOnView has been updated with the presence of the coordinates.

Class GameTable and TestTurnManager and TestGameTable are now consistent with how the players are actually created (with sessions instead of names).