Full ConnectionController implementation + Test + Sonar
The connection controller is now fully implemented and handle the possible disconnection of an user and his reconnection. In case the limit of player has been reached, notify back the user who attempted to connect that the game is full.

Moreover, there is now an event handling to check if the game is alredy started.

Add:
- SetupController which setup the server class after the game setup chosen by the first user who connected.
- MainClass to start the server
- Mockito reference to pom

Update:
- RMISettings with the static parameter to contact the server
- ShootController with some fix because of model alteration
- LoginHandler now handle the amount of time for each action an user have to make his move. It's demanded to it because he have notion of all the user access information in order to handle them properly.

Remove:
- Client and Server class. The launch is demanded at one class, the "MainClass", which based on the args passed will launch as Server or as Client.