ActionState completed, PowerUpState partially implemented
ActionState completed.

PowerUpState has now some methods that can be used from the outside to interact with the user with two simple methods, when the use of a powerup is possible. Detailed explanation in the javadoc.

Deleted the possibility to exit from the game with a command: it is useless because the user can already simply exit from the game clicking the usual cross on the right top side of the window of the command line.