Complete InfoOnView and minor improvements
SquareLM: explanation javadoc made clearer for getPlayers() : they are ids, not PlayerDataLM.

TestPlayer: improved test to be sure mark line and damage line works respecting the respective limits(max 3 marks for each player, max 12 size of damage line).

InfoOnView: completed(the class that manages the UPDATE messages and that shows the map, the players and also contains some information on what the player can see that sometimes may be useful to other parts of the view).