Complete weapon and add new distance calculation
The base effect of all the weapon are now handled.
Railgun and vortex needed a class for their own because they don't have common feature with other weapon.

Fix bug: the check for the movement chosen by the user when using a weapon was not correct. Know, after a couple of more test, should be fixed.

Add the 4 cardinal direction distance calculation

Update the weapon test in order to make them more flexible