Add TypeAdapter for the SquareLM
In order to handle the serialization of light model of the square, a type adapter has been made.

In the main class now the observer are correctly linked to their observable.

Fix a bug in the view where consider the id room wrong to set a param even if it is correct.