Various bug fix + add jar file and respective manifest
Updated the GameTable so that it should set all the cards on the table before the first spawn of the first player (but we still have to verify that actually works as expected).

Players should now be seen as P1, P2 instead of 81, 82 etc.

Removed replicated code in resetDecksIfNecessary() of the TurnManager (and tested).

Added in ToolsView methods to print coordinates for users more easily: printCoordinatesForUser(...).