Bug fix
GeneralMessageObserver: receive the update with the eventType ERROR correctly (without trying to deserialize anything).

LoginState: the loop that keep asking the credentials in case the name was already existing.

PowerUpState: now handle the case of trying to use Newton when you can't and telling the user they can't use it.

SpawnState: the cards just drawn for the spawn are now immediately added to the light model.